# internal/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a chevron remote-signer server",
        "title": "Remote Signer API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://tldrlegal.com/license/mit-license"
        },
        "version": "1.4"
    },
    "basePath": "/remoteSigner",
    "paths": {
        "/agent": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "This signs the request with the key that the specified token is using.",
                "operationId": "agent-proxy-call",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Proxy Token generated with agentAdmin. It is required if running with authentication enabled",
                        "name": "proxyToken",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Target server URL. Defaults to environment variable AGENT_TARGET_URL",
                        "name": "serverUrl",
                        "in": "header"
                    },
                    {
                        "description": "POST Content to send signed to the target server. The message body will be sent to the target server with it's signature in a header field named 'signature'.",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.graphqlQueryExample"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "result of the query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/QuantoError.ErrorObject"
                        }
                    }
                }
            }
        },
        "/agentAdmin": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "This is the GraphQL Endpoint for administration of the agent proxy tokens.",
                "operationId": "agent-proxy-admin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Proxy Token of the admin user. It is required for all calls besides the login",
                        "name": "proxyToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "The JSON content of the graphql query",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.graphqlQueryExample"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "result of the query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/QuantoError.ErrorObject"
                        }
                    }
                }
            }
        },
        "/fieldCipher/cipher": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Field Cipher"
                ],
                "summary": "Encrypt JSON fields to specified gpg keys.",
                "operationId": "field-cipher-cipher",
                "parameters": [
                    {
                        "description": "The encryption parameters",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FieldCipherInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fieldcipher.CipherPacket"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/QuantoError.ErrorObject"
                        }
                    }
                }
            }
        },
        "/fieldCipher/decipher": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Field Cipher"
                ],
                "summary": "Decrypt JSON fields from specified gpg keys.",
                "operationId": "field-cipher-decipher",
                "parameters": [
                    {
                        "description": "The decryption parameters",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FieldDecipherInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fieldcipher.DecipherPacket"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/QuantoError.ErrorObject"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "QuantoError.ErrorObject": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string",
                    "example": "INVALID_FIELD_DATA"
                },
                "errorData": {
                    "type": "object"
                },
                "errorField": {
                    "type": "string",
                    "example": "fieldName"
                },
                "message": {
                    "type": "string",
                    "example": "error example"
                },
                "stackTrace": {
                    "type": "string"
                }
            }
        },
        "fieldcipher.CipherPacket": {
            "type": "object",
            "properties": {
                "encryptedJSON": {
                    "type": "object",
                    "additionalProperties": true
                },
                "encryptedKey": {
                    "type": "string",
                    "example": "wcDMA8HPMfuMKotZAQwAo62NR4snfqbT3S3EBd3xKAJjJuRx42hJU/f+p0eiQvXNuitRuLe0rF0U8YB3ArAhMX1OZ27t/QE7LKDd1T1oY28kUnHzkKzaIBoted7YXveXLRgr5WI1L6impgxlv+88C81Q6h7RqVWG2Vo6+rXdtg7GdK/VEOtJezIlRJ9Od/gBxmGFjtbSzeoQUTXyzN+xPY60PjpX1FXx+gmM1wHGvZjNLUSsMoKE01JtJJQj1kD4MX9nusp0CONzY4oCNptxgFgcSI/AFj7MZJAW9nH4yR+lQrjw+2KeAhWsWebGK4WiZFdxbEkVJ26GSawCTUqvqJJVt3R7N8vEmgNmM5u+QugM9inFQVa8SUTfqdHmpxq/QO+HtOqbsEiBZWHfNIC1muqjEshwpGhvqfajinSkyR2PbzwUgxPneTrGHiV/cG2LdriAy2zUjNSyoXsYqB9sp3gs9KdKg6nh+f0YE4fAwnb91+2B7xJz0wJFm25iAT4VkJCZjWOULVOzAEJv/38C0uAB5JjPc2wU364MKjwj+/iNutXhTHLgbeD44dmp4GfkCmlO9Hh8TM9JZXOLSgd7uODg4nkTH3ngvOIoe2BD4NTlA/sAp+tXZWRSPinNVIvt1MQgi4QcnHl3SNajPDCqZT7gSOTi7B1gegeF7uHuCeA7IaxK4qmlohvhOFAA"
                }
            }
        },
        "fieldcipher.DecipherPacket": {
            "type": "object",
            "properties": {
                "decryptedData": {
                    "type": "object",
                    "additionalProperties": true
                },
                "jsonchanged": {
                    "type": "boolean",
                    "example": false
                },
                "unmatchedFields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fieldcipher.UnmatchedField"
                    }
                }
            }
        },
        "fieldcipher.UnmatchedField": {
            "type": "object",
            "properties": {
                "expected": {
                    "type": "string",
                    "example": "/data/Test/0/test/"
                },
                "got": {
                    "type": "string",
                    "example": "/data/Test/0/name/"
                }
            }
        },
        "models.FieldCipherInput": {
            "type": "object",
            "properties": {
                "json": {
                    "type": "object",
                    "additionalProperties": true
                },
                "keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "0551F452ABE463A4"
                    ]
                },
                "skipFields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.FieldDecipherInput": {
            "type": "object",
            "properties": {
                "encryptedJSON": {
                    "type": "object",
                    "additionalProperties": true
                },
                "encryptedKey": {
                    "type": "string"
                },
                "keyFingerprint": {
                    "type": "string"
                }
            }
        },
...
        "server.graphqlQueryExample": {
            "type": "object",
            "properties": {
                "query": {
                    "type": "string",
                    "example": "query me { User_viewer { me { baseName } } }"
                },
                "variables": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        }
    },
...
}